Add files via upload
Java : In Android when we pick an image, video, or any other type of file from the gallery, documents folder, or capture directly from the camera, we receive a URI object in intent. We have different ways to get a full file path from URI based on content providers. In this article, we will discuss how to get a file path from URI in Android for the coming files picked through Storage, Documents, as well as depending upon the _data field for the MediaStore and other file-based ContentProviders.

Kotlin : When we pick an image, video, audio, or any other file from the gallery or documents directory we receive URI in intent in the onActivityResult(...) method. So we always further required to get a full file path from that URI. If we directly try to get the path through URI.getData() method, it doesn’t serve the purpose as it provides FileManager address but not the full file path present in the external storage. The following utility class will help you to Get Path From URI In Kotlin Android.